mac: cefclient: Fix compile error with Xcode 14.3.1

+++ tests/cefclient/browser/text_input_client_osr_mac.mm
@@ -132,7 +132,7 @@ extern NSString* NSTextInputReplacementRangeAttributeName;
 
   BOOL isAttributedString = [aString isKindOfClass:[NSAttributedString class]];
   NSString* im_text = isAttributedString ? [aString string] : aString;
-  uint32_t length = [im_text length];
+  uint32_t length = static_cast<uint32_t>([im_text length]);
 
   // |markedRange_| will get set in a callback from ImeSetComposition().
   selectedRange_ = newSelRange;
